requirements.txt
simplejson
ecdsa
flask
python-dotenv
pymongo
requests